fix compiler warning

+++ drivers/ble/common.c
@@ -10,12 +10,13 @@ void ble_adv_payload_init(struct ble_adv_payload *buf)
 int ble_adv_payload_add(struct ble_adv_payload *buf, uint8_t type,
 			const void *data, uint8_t data_len)
 {
-	if ((buf->index + data_len + 2/*len+type*/) > sizeof(buf->payload)) {
+	if (((size_t)buf->index + data_len + 2/*len+type*/) >
+			sizeof(buf->payload)) {
 		return -ENOSPC;
 	}
 
 	uint8_t *p = &buf->payload[buf->index];
-	buf->index += data_len + 2;
+	buf->index += (uint8_t)(data_len + 2);
 
 	p[0] = data_len + 1/*type*/;
 	p[1] = type;
